SIM-2873: Fix all the changes from the fable_saga refactor.
* Mainly, we are overriding endpoints now.
