+++ python/mypy.ini
@@ -1,5 +1,6 @@
 [mypy]
-packages = demos, thistle_gulch
+packages = demos, thistle_gulch, tests
+files = *.py
 ;color_output = True
 # set the stubs path
 mypy_path = ./stubs
