+++ python/tests/test_dummy.py
@@ -1,3 +1,2 @@
-class TestDummy:
-    def test_dummy(self):
-        assert 1 == 1
+def test_dummy():
+    assert 1 == 1
